requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-core
twilio
pydantic